Implemented the <Page> system and revamped other systems.
Implemented the <Page> system. (CreatePage, PageGetMD, PageSetMD, PageExpand)

Revamped systems: 'System', 'MemPage' and so on to 'System->buffers:cosaMDBuffers', 'MemPage_Blocks', 'MemPage_Freed' and so on.
Removed the <BlockLinks> system.
Added more OS things to the 'SysInfo' Page, specifically the 'Limits' of resources.

Pages make's things a bit more complicated and a bit inefficient by runtime, but it enables the usage of one Global System for all systems and enables future plans for the Task system.
It also enables a more dynamic and seamless integration of Module-Extensions.
It also removes the unnecessary bloat and complexity of 'Static' and 'Dynamic' compile modes for the bit ranges of said Systems, so no need to worry about creating four different Levels of the same intended Functionality.
(And things are way easier to comprehend now. ._.')

Optimized 'cosaOP*Area' functions somehow EVEN further, compiles just about 20-35 instructions each now in the test environment - If I remember the numbers correctly on the back of my head.

Compiled binary went from 89240Bytes to 69728Bytes - Both binaries compiled with the same flags, about 19.10KB total difference. :p


(I was sick for two weeks and it still feels like my head is burning up.. Anyhow, going to work on the 'StrPage' and 'FilePage' system now, may be adding function names for future systems.. I also need to figure out the generalization of GPU Architectures, maybe I should create a separate module for now that will be merged with the future-to-be 'nCPAA' project in the future?)